Add edit gallery functionality to admin module
The commit introduces an edit functionality to the galleries in the admin module, excluding the 'instax' gallery. This is implemented via adding an "Edit" button for each gallery, which redirects to a new EditGallery component. This allows the admin user to browse and modify the contents of each gallery. The "AddPhoto" component has also been refactored for reusability.