Refactor gallery, admin, posts and locales modules
This commit refactors the gallery, admin, posts and locales modules across both CSS styling and functionality. It introduces an option to show titles for gallery items and makes adjustments to layout styles, particularly to image and font sizes. In navigation, it replaces the use of Button and Link with the more flexible useNavigate hook from react-router-dom. In locales, it adds new translations for situations when no posts or galleries are found.